third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-attachment value 'local'
  assert_equals: expected "local" but got "local, local, local"
[FAIL] Property background-attachment value 'scroll, fixed'
  assert_equals: expected "scroll, fixed" but got "scroll, fixed, scroll"
[FAIL] Property background-attachment value 'local, fixed, scroll, fixed'
  assert_equals: expected "local, fixed, scroll, fixed" but got "local, fixed, scroll"
[FAIL] Property background-clip value 'border-box'
  assert_equals: expected "border-box" but got "border-box, border-box, border-box"
[FAIL] Property background-clip value 'content-box, border-box'
  assert_equals: expected "content-box, border-box" but got "content-box, border-box, content-box"
[FAIL] Property background-clip value 'content-box, border-box, padding-box, content-box'
  assert_equals: expected "content-box, border-box, padding-box, content-box" but got "content-box, border-box, padding-box"
[FAIL] Property background-clip value 'border-area text'
  assert_equals: expected "border-area text" but got "border-area text, border-area text, border-area text"
[FAIL] Property background-clip value 'border-area text, content-box'
  assert_equals: expected "border-area text, content-box" but got "border-area text, content-box, border-area text"
[FAIL] Property background-origin value 'border-box'
  assert_equals: expected "border-box" but got "border-box, border-box, border-box"
[FAIL] Property background-origin value 'content-box, border-box'
  assert_equals: expected "content-box, border-box" but got "content-box, border-box, content-box"
[FAIL] Property background-origin value 'content-box, border-box, padding-box, content-box'
  assert_equals: expected "content-box, border-box, padding-box, content-box" but got "content-box, border-box, padding-box"
[FAIL] Property background-position value '50% 6px'
  assert_equals: expected "50% 6px" but got "50% 6px, 50% 6px, 50% 6px"
[FAIL] Property background-position value '12px 13px, 50% 6px'
  assert_equals: expected "12px 13px, 50% 6px" but got "12px 13px, 50% 6px, 12px 13px"
[FAIL] Property background-position value '12px 13px, 50% 6px, 30px -10px, -7px 8px'
  assert_equals: expected "12px 13px, 50% 6px, 30px -10px, -7px 8px" but got "12px 13px, 50% 6px, 30px -10px"
[FAIL] Property background-position-x value '0.5em'
  assert_equals: expected "20px" but got "20px, 20px, 20px"
[FAIL] Property background-position-x value '-20%, 10px'
  assert_equals: expected "-20%, 10px" but got "-20%, 10px, -20%"
[FAIL] Property background-position-x value 'calc(10px - 0.5em), -20%, right, 15%'
  assert_equals: expected "-10px, -20%, 100%, 15%" but got "-10px, -20%, 100%"
[FAIL] Property background-position-y value '0.5em'
  assert_equals: expected "20px" but got "20px, 20px, 20px"
[FAIL] Property background-position-y value '-20%, 10px'
  assert_equals: expected "-20%, 10px" but got "-20%, 10px, -20%"
[FAIL] Property background-position-y value 'calc(10px - 0.5em), -20%, bottom, 15%'
  assert_equals: expected "-10px, -20%, 100%, 15%" but got "-10px, -20%, 100%"
[FAIL] Property background-repeat value 'round'
  assert_equals: expected "round" but got "round, round, round"
[FAIL] Property background-repeat value 'repeat-x, repeat'
  assert_equals: expected "repeat-x, repeat" but got "repeat-x, repeat, repeat-x"
[FAIL] Property background-repeat value 'repeat-y, round no-repeat, repeat-x, repeat'
  assert_equals: expected "repeat-y, round no-repeat, repeat-x, repeat" but got "repeat-y, round no-repeat, repeat-x"
[FAIL] Property background-size value 'contain'
  assert_equals: expected "contain" but got "contain, contain, contain"
[FAIL] Property background-size value 'auto 1px, 2% 3%'
  assert_equals: expected "auto 1px, 2% 3%" but got "auto 1px, 2% 3%, auto 1px"
[FAIL] Property background-size value 'auto 1px, 2% 3%, contain, 7px 8px'
  assert_equals: expected "auto 1px, 2% 3%, contain, 7px 8px" but got "auto 1px, 2% 3%, contain"
Harness: the test ran to completion.
